requirements.txt
requests==2.31.0
torch==2.1.0
boto3==1.34.11
tqdm==4.65.0
diffusers==0.25.0
toml==0.10.2
omegaconf==2.3.0
transformers==4.37.0
schedule==1.2.1
python-dotenv==1.0.1
peft==0.10.0
openai==1.14.3
accelerate==0.29.1
web3==6.18.0
mnemonic==0.21
prettytable==3.10.0
